+++ tests/testthat/test_dir_recursive/helper_hello.R
@@ -1,0 +1,1 @@
+hello <- function() "Hello World"