+++ tests/testthat/test_dir_recursive/test-empty.R
@@ -1,0 +1,3 @@
+test_that("empty test", NULL)
+
+test_that("empty test with error", stop("Argh"))